format

+++ pclib-db-spring/src/test/java/lu/kbra/pclib/db/base/MoreTypeFactory.java
@@ -3,9 +3,6 @@ package lu.kbra.pclib.db.base;
 import java.lang.reflect.Type;
 import java.util.List;
 
-import lombok.Getter;
-import lombok.NonNull;
-import lombok.RequiredArgsConstructor;
 import lu.kbra.pclib.db.dbms.MySQLDbmsProvider;
 import lu.kbra.pclib.db.dbms.SQLiteDbmsProvider;
 import lu.kbra.pclib.db.domain.column.type.ColumnType;
@@ -14,22 +11,26 @@ import lu.kbra.pclib.db.type.factory.DatabaseTypeFactory;
 import lu.kbra.pclib.db.utils.registry.ColumnTypeFactory;
 import lu.kbra.pclib.db.utils.registry.ColumnTypeRegistry;
 
+import lombok.Getter;
+import lombok.NonNull;
+import lombok.RequiredArgsConstructor;
+
 public class MoreTypeFactory implements DatabaseTypeFactory {
 
 	@RequiredArgsConstructor
-	public class AgeType implements ColumnType<Age, Integer> {
+	public class AgeType implements ColumnType<Age, Long> {
 
 		@Getter
-		private final EncodingType<Integer> encodingType;
+		private final EncodingType<Long> encodingType;
 
 		@Override
-		public @NonNull Age decode(@NonNull Integer value, Type type) {
+		public @NonNull Age decode(@NonNull Long value, Type type) {
 			return new Age(value.byteValue());
 		}
 
 		@Override
-		public @NonNull Integer encode(@NonNull Age value) {
-			return (int) value.value();
+		public @NonNull Long encode(@NonNull Age value) {
+			return (long) value.value();
 		}
 
 	}
@@ -42,7 +43,7 @@ public class MoreTypeFactory implements DatabaseTypeFactory {
 	public void registerColumnTypes(final List<ColumnTypeFactory<?>> typeMap) {
 		ColumnTypeRegistry.registerType(AgeType.class,
 				(clazz, typeHints, etp) -> clazz == Age.class ? ColumnTypeRegistry.TYPE_CATCH_ALL_SCORE : ColumnTypeRegistry.EXCLUDE,
-				(optType, typeHints, etp) -> new AgeType(etp.getTypeFor(Integer.class)), // TODO this should be Byte.class
+				(optType, typeHints, etp) -> new AgeType(etp.getTypeFor(Long.class)), // TODO this should be Byte.class
 				typeMap);
 	}
